NA_Project/requirements.txt
asgiref==3.2.7
Django==3.0.5
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1

requests~=2.24.0
djangorestframework~=3.11.0